px live should support passing in no script name, bug fixes.
Summary:
D7641 changed the old behavior where px live did not need a script name. This was done with the intention of making it more similar to px run to prepare for a refactor of common code between them, but it's actually desired that their behavior differs here. This diff brings it back to the previous behavior.

Additionally, there were a few bugs with px live where arguments were not getting parsed due to existing bugs that these changes exposed, a well as errors getting squashed (also exposed by D7641). This diff causes those errors to be propagated to the user, and also ensures that script arguments in px live are always parsed.

Test Plan: ran various cases

Reviewers: michelle, zasgar, vihang

Reviewed By: michelle

Differential Revision: https://phab.corp.pixielabs.ai/D7666

GitOrigin-RevId: cc44b4b7bc1ab9d6c957ea77f75f1f2060da736f