Reorganize YAML template code
Summary:
as part of the work to have Vizier deploy without the bootstrap update job, the release build will be updated so that it uploads the templated YAMLs to GCS. this is so we don't have to regenerate the template files on the fly whenever the user wants to deploy (this is slow).

the template files are currently generated for our helm charts using `px generate-yamls`. however, i realized that by doing so, this may cause some confusion about whether or not the template changes should be tied to CLI/Vizier releases. since the template is created based off of a specific Vizier release, it should be tied to vizier releases.
to avoid this confusion, this diff unpairs the template generation code from the CLI. the next diff will replace `px generate-yamls` with a separate binary which the release build can execute.

there are no logic changes in this diff--just movement of code to different packages.

Test Plan: ran `px deploy`, confirm it still works.

Reviewers: vihang, zasgar

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D7668

GitOrigin-RevId: 52bc420218a2749d883c2f2f975cdf1accb58b2a