PP-2445: Add a rule that checks for a given IR graph, whether or not its UDFs can successfully execute on a Kelvin (or a PEM)
Summary: This rule verifies that for a given IR graph, that it can successfully execute on a Kelvin (and there is an equivalent rule for PEMs). I think that eventually the UDTF executor logic may be pulled into this rule, but it is no longer invoked in the splitter so I will leave them separate for now. This rule will be used to determine a. if the distributed splitter has split the graph in an appropriate way (no Kelvin UDFs on PEM plans) and b. its static method OperatorUDFsRunOnPEM/Kelvin will be used to determine that a Kelvin-only UDF is cause to split the plan on that operator, even if that operator itself is not blocking.

Test Plan: added new tests.

Reviewers: philkuz, zasgar

Reviewed By: philkuz

JIRA Issues: PP-2445

Differential Revision: https://phab.corp.pixielabs.ai/D7557

GitOrigin-RevId: c49a22fa3f8530cca31f3c4672f1ccaca7569c11