PP-2492: Add some buffer to GRPC output row batch size to account for downstream metadata
Summary: New Relic's 450 node cluster intermittently gets an error that the max NATS size has been reached, which fails px/cluster about 10-20% of the time. We believe that this is due to the fact that query broker adds some extra fields when sending the message over NATS via the proxy (a string table ID). In this diff, some extra "padding" is added to GRPC sink's max row batch size to account for this discrepancy when calculating the max allowed output size in Carnot.

Test Plan: existing

Reviewers: philkuz, zasgar, michelle

Reviewed By: michelle

JIRA Issues: PP-2492

Differential Revision: https://phab.corp.pixielabs.ai/D7670

GitOrigin-RevId: 313639bc879fe9a072437f240ec368462deda647