PX-88 Update vizier build release to generate and upload templates to GCS
Summary:
we are updating vizier deploys so that they no longer need the bootstrap job.
to do so in its current state, the CLI must download the vizier YAMLs, generate the templated version on the fly, fill in the templates, then apply the YAMLs.
there is no need for the CLI to generate the templates on the fly, as this process is also slow.

this diff updates the vizier release build so that it also uploads the templated YAMLs to GCS. this way, the CLI just needs to download the templated YAMLs, fill them in, then apply.
the next diff will add the templates as a new artifact type, so that the template URLs can be fetched through our ArtifactTracker service.

Test Plan: ran rc and checked the files uploaded to GCS

Reviewers: vihang, zasgar

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D7675

GitOrigin-RevId: 7f2ff7a26c51574836d83d3d5ba75b8993e3b32f