Don't include nil VizierInfos when responding to GetVizierInfos
Summary:
Calling `GetVizierInfos` with bad vizierIDS causes us to allocate a slice of length number of ids, but only fill in the valid infos.
The nils in the slice cause a panic on proto marshalling.

Instead use an empty vizierInfo and check vizierID on the caller side to ensure that we have a valid vizier.
(We could also instead respond with fewer infos, or use a map)

Test Plan: Existing tests.

Reviewers: zasgar, michelle

Reviewed By: michelle

Differential Revision: https://phab.corp.pixielabs.ai/D7677

GitOrigin-RevId: 220dc2d8f50dd8d09cbb9382d54bfe55ea574a5a