PX-88 Add template artifacts to available artifacts in release build
Summary:
we are now adding the templated versions of the vizier YAMLs to our release, so that users don't have to generate the templated versions on the fly when deploying.
this is part of the effort to make vizier deploys work without the update job. Helm has already been switched over, next is the regular deploy and --extract_yaml deploys.
next steps:
- update the CLI deploy so that it uses the templated YAMLs
- update update job to use templated YAMLs (mostly for code cleanup and consistency)

Test Plan: built an RC, made a call to the artifact service to get the download link for the templates.

Reviewers: vihang, zasgar

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D7681

GitOrigin-RevId: ce634449ad16fb0333364020c5b7e435b3a2c1f9