Resolve a newer `react-dev-utils` to patch a hypothetical exploit
Summary: https://github.com/advisories/GHSA-5q6m-3h65-w53x - we had this in our top level dependencies (erroneously), so it's been removed there. However, Storybook still depends upon it, so this updates the `resolutions` field at the root to ensure it has the security fix. Storyboard's usage seems not to invoke the vulnerable function, so this is entirely a hypothetical vulnerability for us. Still, it's worth fixing.

Test Plan: N/A. Only Storybook uses this dependency.

Reviewers: michelle, vihang

Reviewed By: michelle

Differential Revision: https://phab.corp.pixielabs.ai/D7682

GitOrigin-RevId: fb178f74e0aac33b7d66e34f7ad73604c77baf82