PX-88 Make etcd operator deployment dependent on etcd-tls secret
Summary:
we are now allowing users to deploy vizier on non-bootstrap mode, but there is a race condition when applying the YAMLs (from extract_yaml or Helm charts).
we have a job which provisions the secrets, such as the etcd-tls secrets. if the etcd operator deployment starts up first, it will complain it doesnt have the secrets and get into a bad state.
instead, we should make the deployment wait for the etcd-tls secrets before it actually starts up.

Test Plan: built rc, deployed using extract YAMLs

Reviewers: zasgar, vihang

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D7683

GitOrigin-RevId: 807df793e191831e0c670745a3cb48b12fd96d41