Cleanup update job to use templated yamls
Summary:
deploy was updated so that it uses the templated YAMLs to deploy vizier. cleaning up the update job so that it does the same, so that we don't have to stick in the annotations/labels while applying.
also reorganized some things in the job so its slightly cleaner

Test Plan: created rc and deployed it. tested update job for updating older version -> newer version.

Reviewers: vihang, vjain, zasgar

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D7689

GitOrigin-RevId: 0d39ea4ed98743ba57b9321bd93272da666ad497