PX-88 `px deploy` should deploy Vizier without update job
Summary:
we are updating our deploy process so that it no longer relies on the update job.
prior work has been done to upload our templated YAMLs to GCS.
this diff updates `px deploy` so that it downloads the templated YAMLs (if available) and fills them out. the deploy process is updated so that all of vizier is deployed, rather than just the bootstrap YAML.

Test Plan: tried many different combinations of deploying old/new versions of Vizier with old/new versions of the CLI, with/without etcd operator, and with/without --extract_yaml

Reviewers: zasgar, vihang, vjain

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D7685

GitOrigin-RevId: ccefe21b5954c238d4b1934f874c654aadfce216