PP-2503: When the query plan string is over 1MB, split it into two result row batches so it doesn't trigger an error in NATS
Summary: We want to be able to analyze the critical path of slow queries, which often run on large cluster. The larger the query, the more likely it is that we will want to analyze its performance. Unfortunately, large queries often result in a query plan larger than 1MB, such as the query plan for px/cluser on New Relic's 450 node cluster. We were unable to even view the query plan for 30s long queries, because the query plan result would be over 1MB, thus triggering a rejection from NATS, since 1MB is its max message size. This diff takes a semi ugly solution of splitting that query plan into two responses. If it is larger than 1MB, we will have to copy and paste each piece of the query plan from the UI in order to view the graph. Ideally, we would have a widget in the UI intended to view the query plan in a nicer way, since most of the online graph visualizers can't even handle that large of a graph anyway, but for now, at least this solution will allow us to get the query plan without causing the entire query to fail.

Test Plan: added unit tests

Reviewers: philkuz, michelle, vihang, zasgar

Reviewed By: philkuz, vihang

JIRA Issues: PP-2503

Differential Revision: https://phab.corp.pixielabs.ai/D7674

GitOrigin-RevId: 774eff286bd345d9aaee3e1af8f4cc5f2b259a96