Update flamegraph colors
Summary:
before we were using the pixie graph colors to color the flamegraph bars. we should do something more similar to brendan gregg and make them various shades of orange/yellow/red. these colors are taken from our Pixie palette

also lessened the opacity of the ancestors, so that it is clearer which bar is selected,

Test Plan:
screenshot:
{F118744}

Reviewers: oazizi, nlanam, jamesbartlett, nserrino

Reviewed By: oazizi

Differential Revision: https://phab.corp.pixielabs.ai/D7709

GitOrigin-RevId: 9ed4b9922d0d12a73b3c1997e79b2673cd4b1649