Fix UI failure when 0 clusters available for an org.
Summary:
UI would fail with a `clusters undefined` error when the user had 0 clusters. We checked whether numClusters was 0, but if there were no clustesr numClusters was actually undefined.
Also had another issue where the UI would try to checks the cluster id of an object that was undefined. Fixed that check as well.

Test Plan: Tested on yarn dev and skaffold.

Reviewers: nlanam, michelle

Reviewed By: michelle

Differential Revision: https://phab.corp.pixielabs.ai/D7732

GitOrigin-RevId: ded6a2fb4a3ed3fe32a2ff4a0c14a60d4240f5f8