Fix cli update check
Summary:
Based on a suggestion in D7661, I changed `GetAvailableVersions` to use
`MustLoadDefaultCredentials` instead of `LoadDefaultCredentials`.

However we check for updates on every invocation of the cli including the ones
where a user is trying to login. Since `MustLoad` exits when no creds exist, this
renders the cli unusable.

This wasn't caught in dev testing becuase a check for `IsDev` short circuits
the `UpdatesAvailable` call.

Until such a time as when the Artifact Service no longer needs auth (PC-796),
we should just optionally try to load creds when trying to get updates.

Test Plan:
Disabled the `IsDev` check. Reproduced the problem. Fixed the problem.
Reenabled the `IsDev` check.

Reviewers: michelle

Reviewed By: michelle

Differential Revision: https://phab.corp.pixielabs.ai/D7735

GitOrigin-RevId: dd2bcdae71e04861c94f2c4714283a87503bf1e9