Remove unnecessary channel closure in query broker
Summary: We were getting a panic in the query broker in a customer cluster where a send would happen on a channel that has already been closed. we don't actually need to close the channel in the places that we are closing it so it was creating unnecessary complexity/error-proneness.

Test Plan: need to run manually.

Reviewers: michelle, vihang, zasgar

Reviewed By: vihang, zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D7743

GitOrigin-RevId: d764fd90677f820e540e1a054965d6b8cc7ad498