PP-2522: Add cloud API for getting Vizier pod list & statuses.
Summary: We currently have debug log as a "debug" service in Vizier. It needs a pod name. The CLI should be able to list those pod names. The UI can list pod names for non-agents, but it's cumbersome to open it in a separate API and it doesn't cover the agent pods. Often we really care about the agent pod logs, but when the cluster is unhealthy, you can't even get those. In the future, we should replace our current approach for getting the control plane pods in the UI with this API, since it will support non-control plane pods as well, and doesn't require storing control plane pod information in the DB.

Test Plan: added unit test.

Reviewers: michelle, vihang, zasgar

Reviewed By: michelle

JIRA Issues: PP-2522

Differential Revision: https://phab.corp.pixielabs.ai/D7770

GitOrigin-RevId: 3cd386e3cbd8fa4fc2b833c860624b5e75e79953