PP-2525: add stub for DebugPods endpoint, support DebugLogs endpoint on non-passthrough clusters
Summary:
Depends on D7770. We want to add "px debug pods" to the CLI so that we can find out the names of pods to query when using px debug log. We also want to subsume the current logic that sends the control plane pod inforation to the UI with this endpoint. Currently px debug log won't work on passthrough clusters because Vizier itself doesn't implement the VizierDebugService. This diff changes that so that the following can happen:
1. px debug log can work on non-passthrough clusters
2. px debug pods can be supported on non-passthrough clusters (coming up next)
3. the debug pods API can subsume the control plane pods being stored in the cluster DB. Then, the UI can use this information instead, including with pod logs if we decide to.

Test Plan: Ran `px debug log` with a non-passthrough cluster without these changes, it failed. Ran `px debug log` with a non-passthrough cluster WITH these changes, it succeeded.

Reviewers: michelle, vihang, zasgar

Reviewed By: michelle

JIRA Issues: PP-2525

Differential Revision: https://phab.corp.pixielabs.ai/D7773

GitOrigin-RevId: d87bb21e04bb7ad587b42935f3919ca426834361