PP-2495 Flamegraph: filter out bars with tiny widths
Summary: some bars have very very small widths which are difficult to click and make it appear as if the flamegraph has weird artifacts. we should just filter them out. when a user clicks into one of the closer ancestors, and the size of the particular trace increases, it will show up.

Test Plan: yarn dev

Reviewers: oazizi, jamesbartlett, nlanam

Reviewed By: oazizi

Subscribers: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D7781

GitOrigin-RevId: fc8567bb19446f233c7a0093a632cd9a7372206d