Limit the number of workers Jest spawns in CI
Summary:
jest seems to be aggresive about trying to detect the number of cores
and spawning workers based on that. https://github.com/facebook/jest/issues/5989

However it seems like jest doesn't handle virtualized environments well. Namely
that it uses `os.cpus()` to detect how many CPUs are available but in an
environment like Kubernetes, the node might have 32 CPUs but only a small number
of them could be available to the current pod.

Instead of trying to get fancy with autodetecting this in our CI environment,
just set it to a simple constant. We don't have enough tests here to over
optimize parallelization. 4x should be good enough.

Test Plan: Jenkins

Reviewers: zasgar, michelle, nlanam

Reviewed By: zasgar, michelle

Differential Revision: https://phab.corp.pixielabs.ai/D7794

GitOrigin-RevId: 391005d74fdfd07519a50a472795eecab588880a