Flamegraph: Make nodeColumn behave like other k8s columns
Summary:
when the pod/container/namespace columns are not specified and cannot be resolved for a stacktrace, we add a bar for the resource which looks like: "namespace: UNKNOWN".
we took a slightly different approach with the nodeColumn, where if the node was not resolved, we just didn't show a bar. for consistency, we should just make the nodes behave like our other k8s resources.

Test Plan: jest test

Reviewers: oazizi, jamesbartlett, zasgar

Reviewed By: zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D7873

GitOrigin-RevId: 2da913c440ee4a239659e3084bacef0ecaa9d93b