Revert "PP-2445, PP-2518: Add limit pushdown to optimizer."
Summary:
This reverts commit eaec9bbfd3a429eb5d22367334ce10bcd34e06d7.
It had a consequence we need to figure out how to account for. Specifically, it would push limits past PEM-only maps. Then, the distributed splitter would break on that limit, and the PEM-only UDF would output empty strings when scheduled on the Kelvin. Need to figure out a holistic way to address optimization that is PEM/Kelvin-aware.

Test Plan: n/a

Reviewers: philkuz, michelle

Reviewed By: michelle

Differential Revision: https://phab.corp.pixielabs.ai/D7986

GitOrigin-RevId: 8c767636e6518505104d0fbce5ced119ddcd66d7