Fix error checking to enable golint errcheck
Summary:
This is one diff in a series of diffs to fix the error checking in our golang code so that we can ultimately enable golint's errcheck.
this lint rule requires errors to be handled, whether that is just a log or returning the error, etc.
doing this in parts.

Test Plan: unit tests pass

Reviewers: zasgar, vihang, nserrino

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D7960

GitOrigin-RevId: 00e13b9778683d67460038586568ed664fc73e9f