Fix error checking to enable golint errcheck in CLI
Summary:
This is one diff in a series of diffs to fix the error checking in our golang code so that we can ultimately enable golint's errcheck.
this lint rule requires errors to be handled, whether that is just a log or returning the error, etc.
this part mainly tackles the CLI.

Test Plan: unit tests pass, also ran CLI

Reviewers: zasgar, vihang, nserrino

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D7963

GitOrigin-RevId: 8e16396798461f2c93eb307e3088eb7e90f94b70