Fix error checking to enable golint errcheck in vizier
Summary:
This is one diff in a series of diffs to fix the error checking in our golang code so that we can ultimately enable golint's errcheck.
this lint rule requires errors to be handled, whether that is just a log or returning the error, etc.
this part mainly tackles Vizier

Test Plan: unit tests pass, also ran vizier in skaffold and made sure it was working as expected

Reviewers: zasgar, vihang, nserrino

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D7965

GitOrigin-RevId: 0e69d5407faea917d21b3043a904f51ea88895d4