Add auth checks for ValidClaims
Summary:
we disabled the ValidClaims checks when we started addding claims that arent attached to a specific user (services/clusters).
now that the structure of those is more fleshed out, we can add the validClaims checks back in for each claim type.

Test Plan: unit test, deployed to staging and tested paths that have service/cluster/user creds

Reviewers: zasgar, vihang

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D7944

GitOrigin-RevId: 9fd39b5dc85589920eba7b54bb88bff7a859bca1