Update jwt validation to check audience
Summary:
we should be using and checking the audience field in our JWTs. the audience should be the consumer of the JWT. for cloud, this is the domain (withpixie.ai/staging.withpixie.dev, etc). for vizier, this should be just "vizier".
for the former, the domain name is now passed as an env variable to all cloud services.
for the latter, this is just hard-coded when creating the env for each service.

this will log the user out of the UI, since their old token is now considered invalid.
in the CLI, if you try to do any operation with the old token, it will complain that the user is unauthenticated, which indicates that they should re-login.
in both cases, the user just needs to login again and everything will work as normal.

Test Plan:
unit tests
ran staging for cloud, and created a new rc for vizier.
validated everything works as expected with:
- new cloud + new vizier
- old cloud + new vizier
- new cloud + old vizier

Reviewers: zasgar, vihang

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D7955

GitOrigin-RevId: 40a3de37b59f73e44d502c76e01a52fa4fafe129