PP-2445: Add pre-split analysis and optimization phases to the distributed splitter
Summary: We noticed when implementing PP-2445 that there were rules we wanted to apply to the IR that were "distributed-aware" (e.g. aware of the difference between pems and kelvins) but not needing to apply to the entire, possibly huge, distributed plan. This diff adds a skeleton for a "pre split" optimizer and analysis phase. This will be for transformations that the distributed splitter wants to make on the plan before it is split. It is used for things that need to pay attention to PEM vs Kelvin but should occur before the distributed split is executed.

Test Plan: will add

Reviewers: philkuz, jamesbartlett, vjain

Reviewed By: philkuz

JIRA Issues: PP-2445

Differential Revision: https://phab.corp.pixielabs.ai/D8006

GitOrigin-RevId: 977df63f6c58f7133d3192aade0c9164e80a35e2