PP-2445: move limit pushdown rule from optimizer to presplit optimizer in order to add PEM-awareness
Summary: Depends on D8006. We previously added a limit pushdown rule, but it was pushing limits past any map - including maps that are PEM-only. This was causing problems, because the distributed splitter splits the plan on limit nodes, meaning that any subsequent node will be placed for execution on Kelvin. Instead of getting rid of limit pushdown, this diff moves it to the pre-split optimizer, and adds PEM awareness so that the limit is not pushed past map nodes that must run on PEM.

Test Plan: added unit

Reviewers: philkuz

Reviewed By: philkuz

JIRA Issues: PP-2445

Differential Revision: https://phab.corp.pixielabs.ai/D8009

GitOrigin-RevId: 430a1279911d6cda53c5ae4bd13bde621674f56d