Use cluster JWT when making cloud requests from Vizier
Summary:
addressing a TODO. we were using service credentials for some vizier initiated requests (VizierConnected). we should acutally use cluster credentials here.
we need to keep the serviceCredentials, since that is used in RegisterVizierDeployment, where the cluster doesn't have an ID yet.

Test Plan: ran on staging and deployed a cluster, confirmed it connected and things run as normal

Reviewers: zasgar, vihang

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D8013

GitOrigin-RevId: 22e02ab8d7149e83f83fcc660afd4d4ba822b07b