[PP-2586] Add loop close for the run() call
Summary:
This is the cryptic message that you get when running `script_with_multiple_tables.py` but not `..._single_tables.py````Error in sys.excepthook:

Original exception was:```

I suspect it's from asyncio because if you override excepthook this starts working.

The diff between the two examples is the `def func` that occurs in multiple tables. So somethign weird with that.

Adding the loop close for `run()` helps with this. I wanted to make sure that users would be responsible for this if they used the async, so I only touched the sync call.

Test Plan: Bug went away for the above situatoins

Reviewers: vihang, zasgar, vjain

Reviewed By: vihang, zasgar

JIRA Issues: PP-2586

Differential Revision: https://phab.corp.pixielabs.ai/D8060

GitOrigin-RevId: efe12e68b1a9ec8af010e74030ffd266f87cfb17