[PP-2253] Move the dynamic tracing library to planner
Summary:
as part of the OSS release, @zasgar wants to create a better separation of components. Notably, we had a misplacement of the dynamic tracing interface - stirling should implement it, but planner be the owner of the interface. This will allow us to implement future dynamic tracepoints with different underlying implementations if we need to in the future.

Core change is the movement of the proto files, every other change is a downstream effect of that.

Test Plan: Passes all test and builds properly. Should pass Jenkins b4 land.

Reviewers: oazizi, vihang, zasgar, yzhao

Reviewed By: oazizi, vihang, yzhao

Subscribers: zasgar

JIRA Issues: PP-2253

Differential Revision: https://phab.corp.pixielabs.ai/D7895

GitOrigin-RevId: 70ee9c060ac46c91881a51dee303497927e2310e