UI: Don't initialize LaunchDarkly unless client id is specified
Summary:
making sure that the UI works without a LaunchDarkly client ID. without this, things will still work but you will get a bunch of ugly errors printed to the console.
our current plan is to completely opensource the UI without modifications--it shouldnt be a surprise to the user that we're using products such as segment/launchdarkly/announcekit/etc. everything should just continue to work properly if none of the configs are specified.
launchdarkly is a bit special, since it can control which features are shown in the app. for now, anything with a flag that needs to be enabled just wont be enabled if they dont have a launchdarkly configured. since everything is open source, if they really wanted the feature they could change the sourcecode to enable it. those cases are fine, this is mostly here for us to control things on withpixie.ai.

Test Plan:
`yarn dev`, tried without a LD_CLIENT_ID, verified there are no LD errors in the console.
without a LD_CLIENT_ID, also added in a flag to see what would happen. it does not error, the flag value is just undefined, which should be OK.

Reviewers: nlanam, zasgar, vihang

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D8090

GitOrigin-RevId: c789548bc8d854788db788d96e45c8313f48ccaa