Move configs with clientIDs to credentials
Summary:
these configs are technically not secret, but we dont want to check them in with the rest of our repo when we open source.
moved them to /credentials/k8s/<env>/configs, which has a special sops rule that allows devs to read them. this is because we need to read these configs when pointing the webpack build to cloud prod.

Test Plan: ran webpack, make sure that vendors like auth0/announcekit/etc still work as expected.

Reviewers: zasgar, vihang, nlanam

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D8095

GitOrigin-RevId: 5c86a6aff923de9259bcfa1fa6ea5b709b9f5101