Implement old TODO for cluster ID being a required field in graphql endpoints
Summary: Cleaning up old TODOs. Our graphql endpoints to fetch cluster and cluster connection info used to not require an ID and we would return the first cluster available. This was back before the UI fully supported multi-cluster, so it was a temporary transition. Now that that UI supports multicluster, we want to require that a cluster ID be passed in for those endpoints.

Test Plan: edited existing

Reviewers: michelle, vihang, philkuz, nlanam

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D8097

GitOrigin-RevId: c88fa1815366fd54d6eb932351e81b69c5d45dab