More efficient BPF map cleanup
Summary:
Experimenting with a new idea to optimize communication back to the kernel.

Instead of using individual syscalls to delete map entries, we can batch them up,
and then send them collectively as a uprobe. Then the uprobe handles the work by
tracing the arguments.

Still evalutating the peformance gains, but if this works, it could be a useful pattern
in a number of instances.

Test Plan: Existing tests + Manual inspection via bpf_trace_printk

Reviewers: yzhao, jps, rcheng

Reviewed By: jps

JIRA Issues: PP-2550

Differential Revision: https://phab.corp.pixielabs.ai/D7791

GitOrigin-RevId: 3ffc9d33154ac5ab6941ad39e1a9c8de7c7e3cfb