Regenerate Yarn lock; fix resulting breakages
Summary:
This also sets up the dependencies of `@pixie-api/samples-api`, which will contain a demo app for the API library. After adding that and deleting the Yarn lock, I ran `yarn install`. This deduplicated a large number of dependencies, but broke some in the process. To resolve that, I updated ESLint and related packages, then fixed the problems created from doing _that_, and regenerated the Yarn lock one final time to make sure everything works.
The reasons for doing this at all:
* We are overdue for several of these updates
* Duplication was making it impossible to resolve conflicts in the build for this diff's followup
Note that the ESLint updates means we updated the AirBnB config, which seems to have gotten stricter. This is a Good Thing, but will cause for some churn for a while as we slowly please the new rules.

Test Plan: Build, smoke test, run unit tests.

Reviewers: #third_party_approvers, michelle, vihang, zasgar

Reviewed By: #third_party_approvers, vihang

JIRA Issues: PC-736, PC-738

Differential Revision: https://phab.corp.pixielabs.ai/D8112

GitOrigin-RevId: 9156cf781676d30e8f36ec6b99c3daab4041df9a