Cleanup prototool invocation
Summary:
Since we import `gogo.proto` but the dependency is fetched by bazel
we end up having to copy the bazel fetched gogo lib as a workaround to run
prototool. Having the file checked in for prototool to use simplifies this
greatly. We will still use the bazel fetched proto for building and hence
care must be taken to avoid divergence, but `gogo.proto` has not changed in
nearly 2.5 years.

Test Plan: `prototool` still works.

Reviewers: zasgar, michelle, #third_party_approvers

Reviewed By: zasgar, michelle, #third_party_approvers

Differential Revision: https://phab.corp.pixielabs.ai/D8123

GitOrigin-RevId: f96ad8ddcb01bb7b6ab066e1930274d07a1d753c