Covert proto break check detector into a linter
Summary:
We currently run prototool as a "unit test" for arcanist and run it
over the entire repo. Instead, let's run it as a linter. This means the checker
will run only on `.proto` files and one at a time instead of the entire repo.
This will also make lint messages show up inline in the diff.

Downgrade from error to warnings since we do sometimes make what would be
considered a breaking change (such as renaming a field) or removing an unused
field. So these are warnings and should be treated as such.

Test Plan:
Made some breaking changes and ran `arc lint`
{F127032}

Reviewers: zasgar, michelle

Reviewed By: michelle

Differential Revision: https://phab.corp.pixielabs.ai/D8128

GitOrigin-RevId: 70c4a66a0f233f39a14b2a8b177236dc157f5818