Replace experimental build checker with a script
Summary:
The linter to check that targets in `//experimental` were marked as
manual was written as a `unit test`. Additionally the linter relied on custom
build rules and checking those rules and as a result missed targets like
`//experimental/stirling/proto_data_model/proto:http_record_pl_go_proto` which
aren't manual since they use `pl_go_proto_library` (due to a lack of a `pl_exp_go_proto_library`).

Instead let's use bazel to query targets that are not marked as manual and convert
this check into a script and regex linter. It's also easier to not have to create
custom targets just to mark things as manual and instead rely on the linter to
enforce the check on normal targets.

Used buildozer to rewrite all rules in `experimental` to have a manual tag and
found a bug in `pl_bpf_cc_resource` where tags weren't being passed down by the
macro and fixed it.

Test Plan:
Unmarked a rule in experimental as manual. Ran `arc lint`.
{F127036}

Reviewers: zasgar, michelle

Reviewed By: zasgar, michelle

Differential Revision: https://phab.corp.pixielabs.ai/D8129

GitOrigin-RevId: 08dc4a93188e178bd12058aff935da0b52f55e79