Replace gazelle tester with linter
Summary:
The gazelle checker was run as an arcanist unit test. It would modify
and autofix the files underneath you but then send out the diff anyway causing
folks to have to `arc diff` again.

Instead, run gazelle in diff mode, and as an arcanist linter and send the diff
to arcanist so that `arc` will ask and apply the patch before sending the diff
out. (Sadly arcanist doesn't accept the patch format, it take a set of original
content plus new content, so generating the diff for arcanist is a bit ...
heavy handed.)

Test Plan:
Make some changes that would make gazelle unhappy. Run `arc lint`
{F127064}

Reviewers: zasgar, michelle

Reviewed By: zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D8131

GitOrigin-RevId: b7523e14821ee959e4f1f3a24c49ea348512f711