Optimization: Protobuf string truncation
Summary:
We spend a lot of time parsing protobufs. One hypothesis is this is because of long strings/bytes in the messages.

We end up truncating the entire message at the end, so it is a wasted effort to render the entire protobuf to a string format, only to have it truncated.

As a middle-ground, truncate all string fields in the protobuf printer.

There could be further optimizations in this library, but let's see how this does.

Test Plan: Created a test for the printer. To evaluate performance savings requires more study.

Reviewers: #stirling, #third_party_approvers, zasgar

Reviewed By: #third_party_approvers, zasgar

Subscribers: yzhao, zasgar

JIRA Issues: PP-2600

Differential Revision: https://phab.corp.pixielabs.ai/D8125

GitOrigin-RevId: ac04cc4922fbfa442459a1aa52e769760984ef07