Fix ci buildables generation
Summary:
The script looks for `_library` `_binary` and `test` target types.
Rename the webpack binary so that it fits that pattern and will get picked up.

Also check to see if a `BUILD.bazel` file changed, and then trigger all targets
in that package.

Test Plan:
Run the script locally with changes inside `src/ui` and a random
`BUILD.bazel` file.

Reviewers: zasgar, michelle, oazizi

Reviewed By: michelle

Differential Revision: https://phab.corp.pixielabs.ai/D8144

GitOrigin-RevId: e9dd5fa411d50999bb7887d2d0b3376d9f7a8e5c