Add TransferDataImpl() to output data to multiple data tables
Summary:
Also hook up it with StirlingImpl, and put it behind a flag:
--stirling_socket_tracer_output_multiple_data_tables (default true)

Tested on my cluster dev-cluster-yzhao, no noticeable changes in performance.
Which is also expected.

The next step is to investigate other ways of iterating ConnTracker objects to exploit the
optimization enabled by this diff.

Test Plan: Jenkins

Reviewers: #stirling, oazizi

Reviewed By: #stirling, oazizi

Subscribers: oazizi

Differential Revision: https://phab.corp.pixielabs.ai/D7952

GitOrigin-RevId: 07a7d63dc9ac25d2223b91a74f5895388ce8520c