Tell prototool to exclude node_modules
Summary:
Prototool walks the entire project on `main` to figure out the original
protos (sigh). Tell it to ignore the hundreds of thousands of files in node_modules.
(Also ignore offline_package_cache which is not quite as large but still large.)

Test Plan: Run prototool on a slower machine.

Reviewers: philkuz

Reviewed By: philkuz

Differential Revision: https://phab.corp.pixielabs.ai/D8160

GitOrigin-RevId: d7938bbd22483c264713f5e80f91adc5b35900f8