Set memory limits on PEM daemonset
Summary:
since we deploy a PEM to each node in a user's cluster, we should make sure that the PEM's memory doesn't get too large.
this diff:
1. updates our templated YAMLs to include a configurable memory limit on the PEM. if no limit is specified, it currently gives a default limit of 2Gi. Clusters that have already been deployed will get this memory limit set by default.
2. updated the CLI deploy so that it now takes a --pem_memory_limit flag, so the user can specify the memory limit they want. they currently cannot specify "no memory limit", since an empty memory limit will be assigned to the default. we can consider adding another flag for disabling memory limits

Test Plan: built rc and deployed, checked the pod to ensure the memory limit was set. deployed to 400+ node test cluster to ensure things are still stable.

Reviewers: zasgar, vihang

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D8195

GitOrigin-RevId: 01306def903981802ccf58247284b0f9c148cb6e