Disble auth for artifactTracker requests
Summary:
we currently require the user to be logged in when making requests to artifactTracker for updating their CLI.
this diff disables the auth checks for the artifactService calls to GetDownloadLink/GetAvailableArtifacts.

It also removes the credentials check for the `px update cli` command.

unfortunately, although this diff removes the auth check in our backend, older CLIs will still try to load the credentials and won't even get to the point where they can make a call to the backend. :( so unfortunately, if a user is not logged into the CLI and can run `px update cli`, they'll have to redownload the CLI from our site.

Test Plan: deployed to staging, deleted `~/.pixie`, and ran `px update cli`.

Reviewers: zasgar, vihang

Reviewed By: vihang

JIRA Issues: PC-796

Differential Revision: https://phab.corp.pixielabs.ai/D8273

GitOrigin-RevId: 9c121a5a218c75da061ac7ae52111ec68e843f25