PP-2445: Add rule that splits out PEM-only UDFs into their own map.
Summary: We have a problem where sometimes UDFs that are PEM-only and Kelvin-only will be scheduled on the same operator. Sometimes, this is the direct result of a user query, for example if they write a filter that uses both types of UDFs in the filter condition. Other times, this is the result of our analyzer, because we merge consecutive maps together in certain conditions, and that merging is not PEM/Kelvin-aware. The consequence of this is that a compiler error will trigger, since these UDFs are mutually exclusive from an execution perspective. This diff introduces a rule which will go in the distributed optimizer in a subsequent diff. The rule splits out PEM-only UDFs into their own maps. This will also enable limit pushdown to push limits past any maps that don't contain any PEM-only UDFs.

Test Plan: added lots of unit tests.

Reviewers: philkuz, jamesbartlett

Reviewed By: philkuz

JIRA Issues: PP-2445

Differential Revision: https://phab.corp.pixielabs.ai/D8153

GitOrigin-RevId: 21d2ac95f75ce3573d6e4f28358d54d6f84399a5